Simplify registration on the example site
Remove the need to create an account as a separate action to logging in
by unifying the Sign up / Log in buttons. Since the example site has no
persistent account DB, this avoids the need to remember to "Sign up" for
an account each time it is run.

This reveals an issue that the `/api/users` call provides an unhelpful
response if a user account associated with the same email address
already exists.