Fix the HYPOTHESIS_SERVICE environment variable
Fix the example HYPOTHESIS_SERVICE environment variable in the README to
be localhost not 127.0.0.1. This is consistent with the setup docs for h
and the client.

If you have the client configured to communicate with h on localhost,
but this publisher test site is configured with HYPOTHESIS_SERVICE set
to 127.0.0.1, then the client will get an "audience (aud) is invalid"
error from h when trying to make grant token request.

By changing HYPOTHESIS_SERVICE to localhost, the grant token request now
succeeds.