Fix crashing when logging in as admin
When you use h's users API to try to create a user that already exists
it sometimes returns a "username already exists" error message and
sometimes an "email already exists".

The publisher test site was failing to recognise the error when it was
"username already exists". Fix it to tolerate either error message.

A planned change to the h service is coming soon that will change the
error message to "username already exists, email already exists". The
code in this commit should work in that case too.