Fix example HYPOTHESIS_AUTHORITY in README
The commands that the README gives for creating the authclient etc in
the h db use the concrete value "partner.org". Change the example
HYPOTHESIS_AUTHORITY environment variable in the same README to use the
same "partner.org" value, rather than a $AUTHORITY placeholder.

I think that HYPOTHESIS_AUTHORITY actually _has_ to match the authority
used when creating the authclient etc in the db (partner.org). This just
makes the docs easier to follow by not making the reader have to figure
that out.