Use a different set of credentials for signing grant tokens
The h service now requires two separate sets of keys for using third
party account. A "client credentials" OAuth client is used for server to
server calls to manage user accounts. A "JWT bearer" OAuth client is
used for signing grant tokens that the client exchanges for access
tokens in order to auto-login as a particular user.